solve merge conflict

+++ src/main/java/com/fptacademy/training/service/dto/ProgramDto.java
@@ -2,6 +2,7 @@ package com.fptacademy.training.service.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 import java.io.Serializable;
@@ -18,7 +19,7 @@ public class ProgramDto implements Serializable {
     private Instant lastModifiedAt;
     private Creator lastModifiedBy;
     private Integer durationInDays;
-    private Integer durationInHours;
+    private Float durationInHours;
     private Boolean activated;
     @Getter
     @AllArgsConstructor
